+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 - Python 3.7 support
 - aiompq==0.11
+- added tests for cabbage.AsyncAmqpRpc.wait_connected()
 
 0.5.3 (2018-07-02)
 ------------------